+++ src/main/java/br/com/invistaix/InvistaIX/repository/EnderecoRepository.java
@@ -8,7 +8,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 
-public interface EnderecoRepository extends JpaRepository<EnderecoModel, Integer> {
+public interface EnderecoRepository extends JpaRepository<EnderecoModel, Long> {
 	@Query("SELECT e FROM EnderecoModel e WHERE e.rua = :rua AND e.numero = :numero AND e.loteamento = :loteamento AND e.cidade = :cidade AND e.estado = :estado AND e.CEP = :CEP")
     Optional<EnderecoModel> findEndereco(@Param("rua") String rua, @Param("numero") int numero, @Param("loteamento") String loteamento, @Param("cidade") String cidade, @Param("estado") String estado, @Param("CEP") String CEP);
 }
